fix(contracts): smart PDF pagination prevents content cutting at page breaks
The contract PDF generator was capturing the entire DOM as one JPEG and
slicing at fixed pixel heights, cutting text, signatures, and tables at
page boundaries. Extracted the estimate's proven smart pagination into a
shared buildPdf utility that measures DOM elements with data-pdf-row
attributes and finds safe break points between content rows. Both
contract and estimate PDF generation now use this shared utility with
page numbers and continuation banners.
